fix : Master - slave 관련 수정
