--- src/main/java/com/smu/smuenip/Infrastructure/config/DataSourceType.java
@@ -1,5 +1,0 @@
-package com.smu.smuenip.Infrastructure.config;
-
-public enum DataSourceType {
-    Master, Slave
-}
