--- src/main/java/com/smu/smuenip/Infrastructure/config/db/MasterDataSourceConfig.java
@@ -1,25 +1,0 @@
-package com.smu.smuenip.Infrastructure.config.db;
-
-import com.zaxxer.hikari.HikariDataSource;
-import javax.sql.DataSource;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-
-@Configuration
-@Slf4j
-public class MasterDataSourceConfig {
-
-    @Primary
-    @Bean(name = "masterDataSource")
-    @ConfigurationProperties(prefix = "spring.datasource.master.hikari")
-    public DataSource masterDataSource() {
-        return DataSourceBuilder.create()
-            .type(HikariDataSource.class)
-            .build();
-    }
-
-}
