+++ src/main/java/com/smu/smuenip/Infrastructure/config/db/ReplicationRoutingDataSource.java
@@ -1,9 +1,9 @@
 package com.smu.smuenip.Infrastructure.config.db;
 
-import com.smu.smuenip.Infrastructure.config.DataSourceType;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
+
 
 @Slf4j
 public class ReplicationRoutingDataSource extends AbstractRoutingDataSource {
@@ -11,7 +11,7 @@ public class ReplicationRoutingDataSource extends AbstractRoutingDataSource {
     @Override
     protected Object determineCurrentLookupKey() {
         return TransactionSynchronizationManager.isCurrentTransactionReadOnly()
-            ? DataSourceType.Slave
-            : DataSourceType.Master;
+            ? "slave"
+            : "master";
     }
 }
