--- src/main/java/com/smu/smuenip/Infrastructure/config/db/RoutingDataSourceConfig.java
@@ -1,42 +1,0 @@
-package com.smu.smuenip.Infrastructure.config.db;
-
-import com.smu.smuenip.Infrastructure.config.DataSourceType;
-import java.util.HashMap;
-import java.util.Map;
-import javax.sql.DataSource;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.DependsOn;
-import org.springframework.jdbc.datasource.LazyConnectionDataSourceProxy;
-
-@Configuration
-@Slf4j
-public class RoutingDataSourceConfig {
-
-    @Bean(name = "routingDataSource")
-    public DataSource routingDataSource(
-        @Qualifier("masterDataSource") final DataSource masterDataSource,
-        @Qualifier("slaveDataSource") final DataSource slaveDataSource
-    ) {
-
-        ReplicationRoutingDataSource routingDataSource = new ReplicationRoutingDataSource();
-
-        Map<Object, Object> dataSourceMap = new HashMap<>();
-
-        dataSourceMap.put(DataSourceType.Master, masterDataSource);
-        dataSourceMap.put(DataSourceType.Slave, slaveDataSource);
-
-        routingDataSource.setTargetDataSources(dataSourceMap);
-        routingDataSource.setDefaultTargetDataSource(masterDataSource);
-
-        return routingDataSource;
-    }
-
-    @Bean
-    @DependsOn("routingDataSource")
-    public LazyConnectionDataSourceProxy dataSource(DataSource routingDataSource) {
-        return new LazyConnectionDataSourceProxy(routingDataSource);
-    }
-}
