--- src/main/java/com/smu/smuenip/Infrastructure/config/db/SlaveDataSourceConfig.java
@@ -1,23 +1,0 @@
-package com.smu.smuenip.Infrastructure.config.db;
-
-import com.zaxxer.hikari.HikariDataSource;
-import javax.sql.DataSource;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@Slf4j
-public class SlaveDataSourceConfig {
-
-    @Bean(name = "slaveDataSource")
-    @ConfigurationProperties(prefix = "spring.datasource.slave.hikari")
-    public DataSource slaveDataSource() {
-        return DataSourceBuilder.create()
-            .type(HikariDataSource.class)
-            .build();
-    }
-
-}
